+++ src-tests/Console/Command/RunTestsCommandTest.php
@@ -9,6 +9,7 @@ use Lmc\Steward\Process\ProcessSet;
 use Lmc\Steward\Process\ProcessSetCreator;
 use Lmc\Steward\Selenium\SeleniumServerAdapter;
 use Symfony\Component\Console\Application;
+use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\Finder\Finder;
@@ -69,12 +70,19 @@ class RunTestsCommandTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider directoryOptionsProvider
      * @param string $directoryOption Passed path type option
-     * @param string $errorBeginning Beginning of error message
+     * @param string $errorBeginning Beginning of exception message
      */
-    public function testShouldStopIfAnyRequiredDirectoryIsNotAccessible($directoryOption, $errorBeginning)
+    public function testShouldThrowExceptionIfAnyRequiredDirectoryIsNotAccessible($directoryOption, $errorBeginning)
     {
         $seleniumAdapterMock = $this->getSeleniumAdapterMock();
         $this->command->setSeleniumAdapter($seleniumAdapterMock);
+
+        $expectedError = sprintf(
+            '%s, make sure it is accessible or define your own path using --%s option',
+            $errorBeginning,
+            $directoryOption
+        );
+        $this->setExpectedException('\RuntimeException', $expectedError);
 
         $this->tester->execute(
             [
@@ -84,15 +92,6 @@ class RunTestsCommandTest extends \PHPUnit_Framework_TestCase
                 '--' . $directoryOption => '/not/accessible'
             ]
         );
-
-        $expectedError = sprintf(
-            '%s, make sure it is accessible or define your own path using --%s option',
-            $errorBeginning,
-            $directoryOption
-        );
-
-        $this->assertContains($expectedError, $this->tester->getDisplay());
-        $this->assertSame(1, $this->tester->getStatusCode());
     }
 
     /**
@@ -104,6 +103,70 @@ class RunTestsCommandTest extends \PHPUnit_Framework_TestCase
             ['tests-dir', 'Path to directory with tests does not exist'],
             ['logs-dir', 'Path to directory with logs does not exist'],
             ['fixtures-dir', 'Base path to directory with fixture files does not exist'],
+        ];
+    }
+
+    public function testShouldOutputAssembledPathsToDirectoriesInDebugMode()
+    {
+        $seleniumAdapterMock = $this->getSeleniumAdapterMock();
+        $this->command->setSeleniumAdapter($seleniumAdapterMock);
+
+        $this->tester->execute(
+            [
+                'command' => $this->command->getName(),
+                'environment' => 'staging',
+                'browser' => 'firefox',
+                '--tests-dir' => __DIR__ . '/Fixtures/DummyTests',
+                '--pattern' => 'NotExisting.foo' // so the test stops execution
+            ],
+            ['verbosity' => OutputInterface::VERBOSITY_DEBUG]
+        );
+
+        $output = $this->tester->getDisplay();
+        $this->assertContains('Base path to fixtures results: ' . realpath(__DIR__) . '/Fixtures/tests', $output);
+        $this->assertContains('Path to logs: ' . realpath(__DIR__) . '/Fixtures/logs', $output);
+        $this->assertContains(' - in directory "' . realpath(__DIR__) . '/Fixtures/DummyTests"', $output);
+        $this->assertSame(1, $this->tester->getStatusCode());
+    }
+
+    /**
+     * @dataProvider browserNameProvider
+     * @param string $browserName
+     * @param bool $shouldThrowException
+     */
+    public function testShouldThrowExceptionIfUnsupportedBrowserSelected($browserName, $shouldThrowException)
+    {
+        if ($shouldThrowException) {
+            $this->setExpectedException('\RuntimeException', 'Browser "' . $browserName . '" is not supported');
+        }
+
+        $seleniumAdapterMock = $this->getSeleniumAdapterMock();
+        $this->command->setSeleniumAdapter($seleniumAdapterMock);
+
+        $this->tester->execute(
+            ['command' => $this->command->getName(), 'environment' => 'prod', 'browser' => $browserName]
+        );
+
+        if (!$shouldThrowException) {
+            $output = $this->tester->getDisplay();
+            $this->assertContains('Browser: ' . strtolower($browserName), $output);
+            $this->assertContains('No testcases found, exiting.', $output);
+        }
+    }
+
+    /**
+     * @return array
+     */
+    public function browserNameProvider()
+    {
+        return [
+            // $browserName, $shouldThrowException
+            'firefox is supported' => ['firefox', false],
+            'chrome is supported' => ['chrome', false],
+            'phantomjs is supported' => ['phantomjs', false],
+            'browser name is case insensitive' => ['FIREFOX', false],
+            'not supported browser' => ['mosaic', true],
+            'unprintable character in browser name' => ['firefox​', true],
         ];
     }